Qt: Fix crash on Windows adding pipe when interface list isn't loaded
The underlying source model for the InterfaceSortFilterModel can
be an InterfaceTreeModel or a InterfaceTreeCacheModel (which do
not inherit from each other), so don't always cast it to an
InterfaceTreeModel pointer and call a member of that on systems
that support remote capture.

It's still impossible to add a pipe when the interface list isn't
loaded on any OS, presumably because of some issues with using
a QIdentityProxyModel when the source model has no entries.
(We don't actually use it as an identity, because the newly added
devices are included too.) However, at least it doesn't crash.

Fix #17928