HTTP/3: Only decode a QPACK encoder stream on the first pass
Store the QPACK encoder stream state, such as the number of insertions
in the current segment and the total so far, and retrieve it on subsequent
passes. Passing the data to the nghttp3_qpack_decoder on subquent
dissections creates confusion and inaccurate results.

We don't have to memdup the buffer unless we're actually doing the
decoding.

We can later use this information in order to defragment QPACK
table insertion instructions that are split across QUIC packets.

Related to #19475