QUIC: Don't desegment and call app handles for 0-length stream segments
Don't bother trying to desegment and calling the subdissector indicated
by the app handle / ALPN for a zero length stream payload segment. It
won't work, and it messes up our methods of determining if desegmentation
is needed.

Revisit this is there ever is a protocol on top of QUIC that must
be called with a zero length payload segment.

Fix #19497