Qt: Don't connect in the .ui file only to disconnect in constructor
Don't connect the QDialogButtonBox accepted() signal to the
dialog's accept() slot in the Qt Creator .ui XML only to
disconnect that connection in the constructor.

Apparently in Qt6 these things are parsed in somewhat different
order, so the disconnection wasn't happening, leading to
destroying the dialog while the FileDialog was still open.

There's a case for moving the Save As actions to ActionRole buttons
(though then the little save icon would have to be set manually.)

Fix #19529