capture: Move pcap failure secondary message to header
Move get_pcap_failure_secondary_error_message from static
in dumpcap to capture-pcap-util-int.h  This means that when
serializing the capabilities, we can just serialize the
open status enum and the open status string, and recover
the secondary status string when deserializing. This also
lets us make the secondary message a pointer to a const
char* instead of doing as much string copying.