epan: Add as much of a text-only subtree length as we can
proto_tree_add_text_valist_internal calls proto_tree_add_text_node,
which calls proto_tree_add_pi() with the internal hf_text_only FT_NONE
item, which calls get_hfi_length which does the adjustment of the
length to be either everything in the tvbuff (if length is -1), or
the minimum of what's in the tvbuff and the length (if length is not -1
and we have the tvbuff.)

This proto_tree_add_text_valist_internal doesn't need to do its own
version of the check. In particular, this means when adding a
text-only subtree with a tvb truncated with capture, the subtree
will be added truncated and the exception will be thrown when
retrieving the first unavailable item, instead of throwing the
exception immediately.

This is the behavior already done when adding FT_PROTOCOL and FT_NONE
items via other methods.

Related to #19544