tshark: Deprecate -G with no argument
We've been planning on removing -G with no argument for
18 years (2f7fd680e20345bbda6731cd041ee9ed0a04690f); start
warning users that it is deprecated.

Single letter options with optional arguments are tricky and
deprecated, see Guideline 12 of the POSIX Utility Syntax Guidelines.
( https://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap12.html )
We have special handling for -G as a result which forces it to the
first argument. -G with no argument can't be mixed with other options,
unlike the other reports. Removing this would allow relaxation of that
restriction.

Related to #17924