Don't include net/if_var.h; it's not present on all *BSDs.
It also doesn't appear to be needed.  If a build fails on some version
of some BSD because it's not being included, *do not* add it back
unconditionally, add it back only for the BSD or BSDs on which it fails.
(NetBSD will not be one of those BSDs, as it doesn't *have* a
net/if_var.h header.)