maxmind_db: Set CLOEXEC on the pipe file descriptors
maxmind_db tries to quit mmdbresolve by closing the pipe that
it uses to write to mmdbresolve, so that mmdbresolve will exit
when it tries to read and the pipe is broken.

However, if mmdbresolve is started, and then later dumpcap is
spawned via fork() and exec*(), either to capture or for capture
stats (sparklines), then the dumpcap process will have open ends
of the pipe, and so mmdbresolve will not exit when disabled
via Preferences->Name Resolution->Enable IP Geolocation.
Instead, Wireshark will hang on waitpid().

Fix the hang by setting FD_CLOEXEC on the pipe's stdin and stdout
file descriptors after spawning the mmdbresolve process, so that
dumpcap does not inherit them.