Revert "Main dialog: Wireshark stops extcap on exit if no packets received yet"
This reverts commit 3aafecb7b95a12e4b1cf52f36c1e1e51e30805ee.

The problem in #17781 was caused by udpdump not flushing after writing
the header, meaning that when no packets were captured, the capture
file wasn't written yet and thus the check in testCaptureFileClose
wouldn't do anything.

This was a workaround for the issue, but the underlying problem
was solved by 9ad1ec1651689a39e86562d083a0794d6fcbc2c2 which ensures
that udpdump and other extcaps flush after writing the pcap header.

The workaround had the downside of requiring the user to enter
the Quit command again (particularly confusing in the case where
closing the capture doesn't result in a dialog, as there's relatively
little notification of what happened outside the icons.)

Fix #19572