iface_lists: Reduce some unnecessary string copying of if_info_t
In scan_local_interfaces(), we copying the information in the
if_info_t from the discovered interface list to the list of all
interfaces. Since we destroy the list of interfaces immediately
afterwards, rather than deep copying the if_info, we can shallow
copy it and remove it from the list.