Qt: Don't offer field completion for literals containing ':'
While ':' is not a character used in fields, it should be added
to the set of token characters in DisplayFilterEdit. It appears
inside byte literals (including addresses), IPv6 literals, and
date and time literals. It is used to separate two literals in
slice notation. It is also now used as a prefix to force ambiguous
values that could be either fields or byte arrays to be interpreted
as byte arrays.

In all those cases, we want to include the ':' in the tokens used
for completion rather than using it as a token separator - since
it is used in literals but not fields, that will prevent field
completion from being offered for literals that include it.
Since there are a number of protocol names that beging with numbers,
and other that begin with hex digits (or even match hex bytes, like
"fc" and "ff"),

The other place in the grammar where ':' is used is to separate
display filter macros from their argument lists in the
${macro:arg1;arg2} format. This change will mean that autocompletion
will no longer be offered for arg1 when typing.

However, it seems likely that that format could be changed to
use all semicolons without harm.

Fix #19499